client/pyproject.toml
[tool.poetry]
name = "myazo"
version = "0.6.0"
description = "Self-hosted, cross-platform Gyazo alternative"
authors = ["Miguel de Moura"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
requests = "^2.24.0"
pyperclip = "^1.8.0"
pillow = "^7.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
